requirements.txt
Django==4.2.7
Pillow==11.0.0
